Allowing clicks on elements that sits below feedback container.
When displaying a Feedback message it used to stay above other elements since it takes 100% width. With this change the feedback container will only take the width it needs to display the message.

TODO: It is not being properly animated, due to some kind of bug with 'display: inline-block' property.