Adding a listing component and also a destroyable listing component.
The ListingView component should be used when you want to reload a static list after adding/removing/updating elements. It renders a LoadingOverlay while loading the list and display an error message if it can't reload the list.

The DestroyableListingView component is a ListingView that can have confirmable destroy links in it. It will properly display a confirmable modal dialog and reload the list after removing an item.