Simplifying loading overlay implementation.
Now it only adds a loading-overlay class to the element, instead of creating an element and resizing it to fit the overlaid element. It comes with some SASS mixins that you can import and use with `dui/mixins`: `overlay-before` that adds a `:before` pseudo-element with a transparent background and `loading-after` that adds a `:after` with a loading centered gif.

MultiSelect now uses loadingOverlay when loading available items.

Added some useful SASS mixins: fade-in, fade-out and ellipsis-overflow.